pyproject.toml
[project]
name = "django-shop-online"
version = "0.1.0"
description = "Depending on the Web Applications of the store online"
authors = [
    {name = "Vadim Bakhtiarov"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "pytz (>=2025.2,<2026.0)",
    "asgiref (>=3.8.1,<4.0.0)",
    "sqlparse (>=0.5.3,<0.6.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-filter (>=25.1,<26.0)",
    "docutils (>=0.21.2,<0.22.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sentry-sdk (>=2.28.0,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "psycopg (>=3.2.8,<4.0.0)",
    "uvicorn (>=0.34.2,<0.35.0)",
    "setuptools (>=80.4.0,<81.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "adrf (>=0.1.9,<0.2.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0", "setuptools>=65.6.3,<81.0.0",]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
ruff = "^0.11.9"
isort = "^6.0.1"
pytest-cov = "^6.1.1"
django-debug-toolbar = "^5.2.0"
pytest-django = "^4.11.1"
pre-commit = "^4.2.0"

[tool.ruff]
line-length = 79
target-version = "py312"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.ruff.lint]
select = ["E", "F", "I", "D"]
ignore = ["D107", "D211", "D212", "E501", "D417"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "mysite.settings"
python_files = "test_*.py"
addopts = "-ra -q"
markers = [
    "django_db: mark the test as using the Django database",
]
